refactor(lib/ecc/field_element): move to module and refactor test
